+++ deno_dist/jsx/dom/index.ts
@@ -1,1 +1,2 @@
 export { render } from './render.ts'
+export { Suspense, ErrorBoundary } from './components.ts'
