+++ deno_dist/jsx/dom/jsx-dev-runtime.ts
@@ -1,0 +1,14 @@
+import type { Props } from '../index.ts'
+
+export const jsxDEV = (tag: string | Function, props: Props, key: string | undefined) => {
+  const children = 'children' in props ? props.children : []
+  delete props['children']
+  return {
+    tag,
+    props,
+    key,
+    children: Array.isArray(children) ? children : [children],
+  }
+}
+
+export const Fragment = (props: Record<string, unknown>) => jsxDEV('', props, undefined)
